requirements.txt
pytest==7.4.3
huggingface-hub<0.17.0
sentence-transformers==2.2.2
chromadb==0.4.22
openai==1.3.5
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
pypdf==3.17.1
python-docx==0.8.11
tiktoken==0.5.1
langchain==0.0.350
ftfy==6.1.1
